Refactoring Services to Classes
Refactored the Quote and Database services to be classes so they can have class properties instead of global variables. There is a single method in the QuoteService that depends on the DBService being passed in which I'm not sure I love. Further work to do. Ethan said he's gonna build a configuration U.I. for quote bot